jobmaster/requirements.txt
fastapi
uvicorn
celery
redis
python-dotenv
psycopg2-binary